requirements.txt
pyspark==3.4.1
kafka-python==2.0.2
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
tensorflow==2.12.0
pandas==2.0.1
dash==2.9.3
plotly==5.14.1
numpy==1.24.3
pymongo==4.3.3